new: [hack.lu 2020] initial template using kaobook
The template was selected because it's somehow similar to the Tufte
template where a column is kept for adding figures, notes and additional
information. This allows the book to be more readable and accessible.
The template is also functional with large diagram or pictures which
can help us to have a more artistic feeling for the published book.
The class is standard and do not use too much obscure tricks to make
it works. This class should allow us to modify the final layout for
publishing (especially on page spacing).

This repository will include the template for the author.

In the final step, we will have a directory with each chapters and
a main LaTeX file for the book.